+++ core/libs/System/cmtkMemory.cxx
@@ -57,7 +57,7 @@ GetNextPowerOfTwo( size_t k )
     return 1;
   
   k--;
-  for (int i=1; i<sizeof(size_t)*CHAR_BIT; i<<=1)
+  for (size_t i=1; i<sizeof(size_t)*CHAR_BIT; i<<=1)
     k = k | k >> i;
 
   return k+1;
